recommender/requirements.txt
psycopg2-binary
mwclient
mwparserfromhell
unidecode
langid